+++ spec/knossos/algorithm/binary_tree_spec.rb
@@ -23,11 +23,4 @@ RSpec.describe Knossos::Algorithm::BinaryTree do
     end
   end
 
-  describe '#carve' do
-    let(:carver) { described_class.new }
-
-    it 'may be called with an optional, non-keyword argument [seed]' do
-      expect(carver).to respond_to(:carve).with(0..1).arguments.and_keywords(:grid)
-    end
-  end
 end